opendistro-elasticsearch: always generate config
Prior to this change, elasticsearch configuration was only performed if the
elasticsearch package is installed in the same playbook run. This makes it
challenging to iterate on failures, since many important tasks may never get to
run. It also prevents modifying the configuration, or working on systems that
already have the package installed.